refactor(baseplate): split BaseplatePanel into section components
BaseplatePanel.tsx was a single 914-line component mixing three stores,
~20 handlers, free-function math, and five UI sections. Each section is
now its own component reading the stores directly:

- DimensionsSection: sync toggle, steppers, fractional edges, editable mm
  summary, padding schematic + margin fill/detach controls
- BaseSection: connectors/magnets/solid floor/corner radius, self-gating
  on the split/stack state (returns null when it would render empty)
- PhysicalUnitsSection: grid unit, print bed, nozzle, build height, with
  its own expand state + help-jump listener
- panelState.ts: plain getState-based updateBaseplateParams/Param
  mutators plus useBaseplatePanelDerived(), the single home for the
  shape/stack gating flags both sections depend on, so the sections
  cannot drift on how 'shaped' or 'stacked' is decided
- The parent shrinks to composition: sections + StackPrintSection +
  SplitViewStrip + reset dialog (1008 -> ~120 lines; max-lines cleared)
- Each new section gets a sibling test against the real stores

Logic and JSX moved verbatim except one explicit undefined guard in
DimensionsSection where the extracted hook boundary hides the aliased
narrowing the old single scope provided (runtime-equivalent).
