+++ src/features/baseplate/components/BaseplatePanel/BaseplatePanel.tsx
@@ -4,162 +4,57 @@
  * Top-to-bottom information hierarchy:
  * 1. Dimensions: sync-with-layout toggle + grid steppers (the unit readout)
  *    + click-to-edit mm summary + spatial padding schematic. Single unified section.
- * 2. Base: magnet holes, dovetails (when split), corner radius
- * 3. Print Settings: grid unit, print bed size (rarely changed)
- * 4. Split pieces mini-map (only when baseplate is split across print beds)
+ * 2. Stack for printing (experimental) — placed above Base because enabling it
+ *    strips and hides the Base controls below.
+ * 3. Base: magnet holes, dovetails (when split), corner radius
+ * 4. Physical Units: grid unit, print bed size (rarely changed)
+ * 5. Split pieces mini-map (only when baseplate is split across print beds)
+ *
+ * Each section is its own component reading the stores directly; shared
+ * derivations (shape/stack gating) live in `panelState.ts`.
  */
 
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { useShallow } from 'zustand/react/shallow';
 import { useLayoutStore } from '@/core/store/layout';
-import { useSettingsStore } from '@/core/store/settings';
 import { useToastStore } from '@/core/store/toast';
-import { FractionalEdgeToggle } from '@/shared/components/FractionalEdgeToggle';
-import {
-  DEFAULT_BASEPLATE_PARAMS,
-  CONSTRAINTS,
-  MARGIN_MIN_DETACH_MM,
-  SOLID_FLOOR_DEFAULT_MM,
-  SOLID_FLOOR_MIN_MM,
-  SOLID_FLOOR_MAX_MM,
-} from '@/core/constants';
-import { PRINT_SETTINGS_CONSTRAINTS } from '@/shared/printSettings';
-import { NOZZLE_BASELINE } from '@/shared/printSettings/connectorScaling';
-import { useHalfGridModeStore } from '@/core/store/halfGridMode';
-import { Checkbox } from '@/design-system/Checkbox/Checkbox';
-import { RulerIcon, RotateCcwIcon } from '@/design-system/Icon';
+import { DEFAULT_BASEPLATE_PARAMS } from '@/core/constants';
+import { RotateCcwIcon } from '@/design-system/Icon';
 import { useTranslation } from '@/i18n';
-import { StickyGroupHeader } from '@/shared/components/StickyGroupHeader';
-import { SettingsRow } from '@/shared/components/SettingsRow';
-import { DeferredNumberInput } from '@/shared/components/DeferredNumberInput';
-import { PrintBedInput } from '@/shared/components/PrintBedInput';
-import { FeatureToggle } from '@/shared/components/FeatureToggle';
-import { SliderInput, Button, ConfirmDialog, CheckboxRow, SegmentedControl } from '@/design-system';
+import { Button, ConfirmDialog } from '@/design-system';
 import { UserDock } from '@/shared/components/UserDock';
 import { AttributionFooter } from '@/shared/components/AttributionFooter';
 import { useFeatureFlag } from '@/shared/hooks/useFeatureFlag';
 import { HelpTargetMarker } from '@/shared/help/HelpTargetMarker';
-import { helpJumpEventName } from '@/shared/help/helpJumpDispatcher';
 import { useBaseplatePageStore } from '../../store/baseplatePageStore';
-import { EditableDimensions } from '@/shared/components/EditableDimensions';
-import { PaddingSchematic } from './PaddingSchematic';
 import { SplitViewStrip } from './SplitViewStrip';
-import { CornerRadiusControl } from './CornerRadiusControl';
-import { GridDimensionStepper } from './GridDimensionStepper';
-import { ConnectorSampleButton } from './ConnectorSampleButton';
 import { StackPrintSection } from './StackPrintSection';
-import { ConnectorPicker } from './ConnectorPicker';
-import type { ConnectorChoice } from './ConnectorPicker';
-import { resolveOverTileStatus } from '../../utils/overTileStatus';
-import { maxCornerRadiusMm } from '../../utils/buildFullParams';
+import { DimensionsSection } from './DimensionsSection';
+import { BaseSection } from './BaseSection';
+import { PhysicalUnitsSection } from './PhysicalUnitsSection';
 import type { StackPrintParams } from '@/core/types';
-import { PADDING_MAX } from '../PaddingStepper';
-import { Stepper } from '@/design-system/Stepper';
-import {
-  CONNECTOR_FIT_OFFSET_MIN,
-  CONNECTOR_FIT_OFFSET_MAX,
-  CONNECTOR_FIT_OFFSET_STEP,
-} from '@/shared/constants/connectors';
-import type { StoredBaseplateParams } from '@/core/types';
-import { gridUnits, mm, effectiveGridUnitMmY } from '@/core/types';
-import { isSeamConnectorStyle } from '@/shared/types/bin';
-import { cornerCutsMatchVertices } from '@/shared/utils/cornerCutOutline';
-import { padOutline } from '@/shared/utils/padOutline';
-
-/** How the drawer-fit padding margin is filled. */
-type MarginFillMode = 'solid' | 'tile' | 'halfGrid';
-
-/** How the sub-21mm leftover after half-grid packing is rendered. */
-type LeftoverMode = 'grid' | 'solid';
-
-/** Snap a connector fit offset to its step and clamp to the allowed range,
- * absorbing IEEE-754 drift from repeated ±0.05 button clicks. */
-function snapConnectorFitOffset(value: number): number {
-  const snapped = Math.round(value / CONNECTOR_FIT_OFFSET_STEP) * CONNECTOR_FIT_OFFSET_STEP;
-  const clamped = Math.max(CONNECTOR_FIT_OFFSET_MIN, Math.min(CONNECTOR_FIT_OFFSET_MAX, snapped));
-  // Round to 2dp so values like 0.30000000000000004 don't leak into the UI/cache.
-  return Math.round(clamped * 100) / 100;
-}
-
-/** Signed label for the connector fit offset, e.g. "+0.05", "0". The "mm" unit is rendered outside the stepper. */
-function formatConnectorFitOffset(value: number): string {
-  if (value === 0) return '0';
-  const sign = value > 0 ? '+' : '−'; // U+2212 minus for typographic consistency
-  return `${sign}${Math.abs(value)}`;
-}
+import { updateBaseplateParams } from './panelState';
 
 export function BaseplatePanel() {
   const t = useTranslation();
   const cloudSyncEnabled = useFeatureFlag('cloud_sync');
 
-  const {
-    drawerWidth,
-    drawerDepth,
-    drawerOutline,
-    drawerFractionalEdgeX,
-    drawerFractionalEdgeY,
-    gridUnitMm,
-    gridUnitMmY,
-    magnetAnchor,
-    printBedSize,
-    printBedDepth,
-    baseplateParams,
-  } = useLayoutStore(
-    useShallow((state) => ({
-      drawerWidth: state.layout.drawer.width,
-      drawerDepth: state.layout.drawer.depth,
-      drawerOutline: state.layout.drawer.outline,
-      drawerFractionalEdgeX: state.layout.drawer.fractionalEdgeX ?? 'end',
-      drawerFractionalEdgeY: state.layout.drawer.fractionalEdgeY ?? 'end',
-      gridUnitMm: state.layout.gridUnitMm,
-      gridUnitMmY: effectiveGridUnitMmY(state.layout),
-      magnetAnchor: state.layout.magnetAnchor ?? 'edge',
-      printBedSize: state.layout.printBedSize,
-      printBedDepth: state.layout.printBedDepth,
-      baseplateParams: state.layout.baseplateParams ?? DEFAULT_BASEPLATE_PARAMS,
-    }))
+  const stackPrint = useLayoutStore(
+    (state) => (state.layout.baseplateParams ?? DEFAULT_BASEPLATE_PARAMS).stackPrint
   );
+  const printBedSize = useLayoutStore((state) => state.layout.printBedSize);
 
-  const { tiling, hoveredPieceLabel, selectedPieceLabel, exportFormat } = useBaseplatePageStore(
+  const { tiling, hoveredPieceLabel, selectedPieceLabel } = useBaseplatePageStore(
     useShallow((s) => ({
       tiling: s.tiling,
       hoveredPieceLabel: s.hoveredPieceLabel,
       selectedPieceLabel: s.selectedPieceLabel,
-      exportFormat: s.exportFileNameConfig.format,
     }))
   );
   const setHoveredPieceLabel = useBaseplatePageStore((s) => s.setHoveredPieceLabel);
   const setSelectedPieceLabel = useBaseplatePageStore((s) => s.setSelectedPieceLabel);
 
-  const nozzleSizeMm = useSettingsStore((s) => s.settings.printSettings.nozzleSizeMm);
-  const handleNozzleChange = useCallback((value: number) => {
-    const current = useSettingsStore.getState().settings.printSettings;
-    useSettingsStore.getState().updateSetting('printSettings', { ...current, nozzleSizeMm: value });
-  }, []);
-  const maxPrintHeightMm = useSettingsStore((s) => s.settings.printSettings.maxPrintHeightMm);
-  const handleMaxHeightChange = useCallback((value: number) => {
-    const current = useSettingsStore.getState().settings.printSettings;
-    useSettingsStore
-      .getState()
-      .updateSetting('printSettings', { ...current, maxPrintHeightMm: value });
-  }, []);
-
-  const updateParams = useCallback((patch: Partial<StoredBaseplateParams>) => {
-    const current = useLayoutStore.getState().layout.baseplateParams ?? DEFAULT_BASEPLATE_PARAMS;
-    useLayoutStore.getState().setBaseplateParams({ ...current, ...patch });
-  }, []);
-
-  const updateParam = useCallback(
-    <K extends keyof StoredBaseplateParams>(key: K, value: StoredBaseplateParams[K]) => {
-      updateParams({ [key]: value });
-    },
-    [updateParams]
-  );
-
-  const halfGridMode = useHalfGridModeStore((s) => s.halfGridMode);
-  const [printSettingsExpanded, setPrintSettingsExpanded] = useState(true);
   const [resetConfirmOpen, setResetConfirmOpen] = useState(false);
-
   const handleReset = useCallback(() => {
     useLayoutStore.getState().setBaseplateParams({ ...DEFAULT_BASEPLATE_PARAMS });
     useToastStore.getState().addToast(t('toast.baseplateReset'), 'success', 3000);
@@ -170,798 +65,19 @@ export function BaseplatePanel() {
   // mutating stored params — so the user's connector settings return intact
   // when stacking is turned off. The connector controls are hidden meanwhile.
   const setStackPrint = useCallback(
-    (next: StackPrintParams | undefined) => updateParams({ stackPrint: next }),
-    [updateParams]
-  );
-  // STEP never stacks, so the controls stacking would strip (magnets, corner
-  // rounding) stay live for STEP exports — mirror BaseplatePage's stackEnabled.
-  const stackEnabled =
-    baseplateParams.stackPrint?.enabled === true && (exportFormat ?? 'stl') !== 'step';
-
-  useEffect(() => {
-    const handler = () => setPrintSettingsExpanded(true);
-    const eventName = helpJumpEventName('baseplate:print-settings');
-    window.addEventListener(eventName, handler);
-    return () => window.removeEventListener(eventName, handler);
-  }, []);
-
-  const synced = baseplateParams.syncWithLayout !== false;
-  const effectiveWidth = synced ? drawerWidth : (baseplateParams.baseplateWidth ?? drawerWidth);
-  const effectiveDepth = synced ? drawerDepth : (baseplateParams.baseplateDepth ?? drawerDepth);
-
-  const handleSyncToggle = useCallback(
-    (checked: boolean) => {
-      const layoutState = useLayoutStore.getState().layout;
-      const current = layoutState.baseplateParams ?? DEFAULT_BASEPLATE_PARAMS;
-      if (checked) {
-        useLayoutStore.getState().setBaseplateParams({ ...current, syncWithLayout: true });
-      } else {
-        useLayoutStore.getState().setBaseplateParams({
-          ...current,
-          syncWithLayout: false,
-          baseplateWidth: drawerWidth,
-          baseplateDepth: drawerDepth,
-          fractionalEdgeX: layoutState.drawer.fractionalEdgeX ?? 'end',
-          fractionalEdgeY: layoutState.drawer.fractionalEdgeY ?? 'end',
-        });
-      }
-    },
-    [drawerWidth, drawerDepth]
-  );
-
-  const handleFractionalEdgeChange = useCallback(
-    (axis: 'x' | 'y', edge: 'start' | 'end') => {
-      if (baseplateParams.syncWithLayout !== false) {
-        useLayoutStore
-          .getState()
-          .updateDrawer(axis === 'x' ? { fractionalEdgeX: edge } : { fractionalEdgeY: edge });
-      } else {
-        updateParam(axis === 'x' ? 'fractionalEdgeX' : 'fractionalEdgeY', edge);
-      }
-    },
-    [baseplateParams.syncWithLayout, updateParam]
-  );
-
-  const gridWidthMm = effectiveWidth * gridUnitMm;
-  // Depth uses the non-square Y pitch, so the panel's dimension readout and the
-  // corner-shape detection match the actual plate (equal to X for a square grid).
-  const gridDepthMm = effectiveDepth * gridUnitMmY;
-
-  const totalWidthMm = gridWidthMm + baseplateParams.paddingLeft + baseplateParams.paddingRight;
-  const totalDepthMm = gridDepthMm + baseplateParams.paddingFront + baseplateParams.paddingBack;
-  const hasPadding =
-    baseplateParams.paddingLeft > 0 ||
-    baseplateParams.paddingRight > 0 ||
-    baseplateParams.paddingFront > 0 ||
-    baseplateParams.paddingBack > 0;
-  const overTileStatus = resolveOverTileStatus(baseplateParams);
-  // Derive the toggle from the STORED flags, not a canOverTile-clamped mode, so
-  // the user can still turn fill OFF when padding temporarily shrinks below the
-  // tile threshold — disabling the whole switch there would strand an enabled
-  // flag that silently re-applies once padding grows again.
-  const fillOn = baseplateParams.overTile === true;
-  const halfGridOn = fillOn && baseplateParams.overTileHalfGrid === true;
-  const setMarginFillMode = useCallback(
-    (mode: MarginFillMode) => {
-      updateParams({
-        overTile: mode !== 'solid',
-        overTileHalfGrid: mode === 'halfGrid' ? true : undefined,
-      });
-    },
-    [updateParams]
-  );
-  const leftoverMode: LeftoverMode =
-    halfGridOn && baseplateParams.overTileHalfGridSolidLeftover === true ? 'solid' : 'grid';
-  const setLeftoverMode = useCallback(
-    (mode: LeftoverMode) => {
-      // Store undefined (not false) for the default Grid so identical geometry
-      // keeps one serialized/cache identity — matches overTileHalfGrid above.
-      updateParam('overTileHalfGridSolidLeftover', mode === 'solid' ? true : undefined);
-    },
-    [updateParam]
-  );
-
-  // Mirrors buildFullParams's outline gate. Corner-cut shapes compose with
-  // padding (the cuts are re-inscribed on the padded rectangle), so padding
-  // stays live and only corner rounding + detached margins hide. Painted/pen/
-  // trace shapes have no parametric resize: they subsume padding too, so all
-  // those controls hide behind the notice. Stacking wins over the shape
-  // (uniform rectangular tiles), so the stack section stays functional — but
-  // via the export-format-aware `stackEnabled`: STEP clears stackPrint before
-  // buildFullParams, so a STEP export of a stacked shaped drawer IS shaped and
-  // the panel must say so.
-  const outlineActive = drawerOutline !== undefined && synced && !stackEnabled;
-  const cornerShaped =
-    outlineActive &&
-    drawerOutline.authoring?.kind === 'corners' &&
-    drawerOutline.authoring.corners !== undefined &&
-    cornerCutsMatchVertices(
-      drawerOutline.vertices,
-      gridWidthMm,
-      gridDepthMm,
-      drawerOutline.authoring.corners
-    );
-  // Every freeform shape (painted cells, traced footprint, pen shape) composes
-  // with padding edge-by-edge in buildFullParams, so its padding controls stay
-  // live — only corner rounding and detached margins (both outline-unaware) hide.
-  const freeformShaped = outlineActive && !cornerShaped;
-  // Mirror buildFullParams: the shape composes only while the padding stays
-  // valid. Padding large enough to fold the loop (a collapsed notch/slot) makes
-  // the resolver drop it, so the panel must say the padding is ignored rather
-  // than present it as applied.
-  const shapedPaddingComposes =
-    !freeformShaped ||
-    padOutline(drawerOutline, {
-      left: baseplateParams.paddingLeft,
-      right: baseplateParams.paddingRight,
-      front: baseplateParams.paddingFront,
-      back: baseplateParams.paddingBack,
-    }) !== null;
-  const canDetach =
-    baseplateParams.paddingLeft >= MARGIN_MIN_DETACH_MM ||
-    baseplateParams.paddingRight >= MARGIN_MIN_DETACH_MM ||
-    baseplateParams.paddingFront >= MARGIN_MIN_DETACH_MM ||
-    baseplateParams.paddingBack >= MARGIN_MIN_DETACH_MM;
-  const detachStored = baseplateParams.detachMargins === true;
-  const marginConnectorStored = baseplateParams.detachMarginConnector === true;
-  // The seam connector reuses the body's tongue/groove; snapClip/dovetailKey
-  // seams would need a separate clip part, so they stay friction-fit (#2414).
-  // `undefined` is the stored default for dovetail, so it counts.
-  const marginConnectorStyleOk = isSeamConnectorStyle(baseplateParams.connectorStyle);
-
-  const hasFractionalWidth = effectiveWidth % 1 !== 0;
-  const hasFractionalDepth = effectiveDepth % 1 !== 0;
-  const fractionalEdgeX = synced
-    ? drawerFractionalEdgeX
-    : (baseplateParams.fractionalEdgeX ?? 'end');
-  const fractionalEdgeY = synced
-    ? drawerFractionalEdgeY
-    : (baseplateParams.fractionalEdgeY ?? 'end');
-
-  const minMm = CONSTRAINTS.GRID_MIN * gridUnitMm;
-  const maxMm = CONSTRAINTS.GRID_MAX * gridUnitMm + PADDING_MAX * 2;
-
-  /**
-   * When the user enters target mm dimensions:
-   * 1. Snap down to the largest valid grid size that fits (half-unit or whole-unit)
-   * 2. Distribute remaining mm evenly as padding (left=right, front=back)
-   * 3. Auto-uncheck "Sync with layout"
-   *
-   * Exception: a synced corner-cut shaped plate keeps sync AND its shape —
-   * the grid is fixed by the layout, so the whole remainder becomes padding
-   * (the basket workflow: measure, type dims, rounded shape persists).
-   */
-  const handleDimensionCommit = useCallback(
-    (targetWidthMm: number, targetDepthMm: number) => {
-      const cornerShapeState = useLayoutStore.getState().layout;
-      const cornerShapeParams = cornerShapeState.baseplateParams ?? DEFAULT_BASEPLATE_PARAMS;
-      const outline = cornerShapeState.drawer.outline;
-      const cuts = outline?.authoring?.kind === 'corners' ? outline.authoring.corners : undefined;
-      // Derive entirely from the fresh store read — mixing in the captured
-      // gridUnitMm could compute padding against a stale drawer size.
-      const drawerWidthMm = cornerShapeState.drawer.width * cornerShapeState.gridUnitMm;
-      const drawerDepthMm = cornerShapeState.drawer.depth * effectiveGridUnitMmY(cornerShapeState);
-      if (
-        cornerShapeParams.syncWithLayout !== false &&
-        outline !== undefined &&
-        cuts !== undefined &&
-        cornerCutsMatchVertices(outline.vertices, drawerWidthMm, drawerDepthMm, cuts)
-      ) {
-        const halfPad = (targetMm: number, gridMm: number): number =>
-          Math.min(PADDING_MAX, Math.max(0, Math.floor(((targetMm - gridMm) / 2) * 100) / 100));
-        const padW = halfPad(targetWidthMm, drawerWidthMm);
-        const padD = halfPad(targetDepthMm, drawerDepthMm);
-        useLayoutStore.getState().setBaseplateParams({
-          ...cornerShapeParams,
-          paddingLeft: mm(padW),
-          paddingRight: mm(padW),
-          paddingFront: mm(padD),
-          paddingBack: mm(padD),
-        });
-        return;
-      }
-
-      const step = halfGridMode ? 0.5 : 1;
-
-      const rawWidthUnits = targetWidthMm / gridUnitMm;
-      const rawDepthUnits = targetDepthMm / gridUnitMmY;
-
-      // Floor so the grid never exceeds the target — remainder becomes positive padding
-      const snappedWidth = Math.max(
-        CONSTRAINTS.GRID_MIN,
-        Math.min(CONSTRAINTS.GRID_MAX, Math.floor(rawWidthUnits / step) * step)
-      );
-      const snappedDepth = Math.max(
-        CONSTRAINTS.GRID_MIN,
-        Math.min(CONSTRAINTS.GRID_MAX, Math.floor(rawDepthUnits / step) * step)
-      );
-
-      const remainderWidth = Math.max(0, targetWidthMm - snappedWidth * gridUnitMm);
-      const remainderDepth = Math.max(0, targetDepthMm - snappedDepth * gridUnitMmY);
-
-      const halfPadWidth = Math.floor((remainderWidth / 2) * 100) / 100;
-      const halfPadDepth = Math.floor((remainderDepth / 2) * 100) / 100;
-
-      const layoutState = useLayoutStore.getState().layout;
-      const current = layoutState.baseplateParams ?? DEFAULT_BASEPLATE_PARAMS;
-      useLayoutStore.getState().setBaseplateParams({
-        ...current,
-        syncWithLayout: false,
-        baseplateWidth: gridUnits(snappedWidth),
-        baseplateDepth: gridUnits(snappedDepth),
-        fractionalEdgeX: layoutState.drawer.fractionalEdgeX ?? 'end',
-        fractionalEdgeY: layoutState.drawer.fractionalEdgeY ?? 'end',
-        paddingLeft: mm(halfPadWidth),
-        paddingRight: mm(halfPadWidth),
-        paddingFront: mm(halfPadDepth),
-        paddingBack: mm(halfPadDepth),
-      });
-    },
-    [gridUnitMm, gridUnitMmY, halfGridMode]
+    (next: StackPrintParams | undefined) => updateBaseplateParams({ stackPrint: next }),
+    []
   );
 
   return (
     <div className="flex h-full flex-col">
       <div className="flex-1 overflow-y-auto scrollbar-thin">
-        {/* 1. Dimensions — sync toggle + grid steppers (steppers = units readout) + mm summary */}
-        <StickyGroupHeader title={t('baseplate.sectionDimensions')}>
-          <div className="space-y-3 px-4 py-3">
-            <Checkbox
-              checked={synced}
-              onChange={handleSyncToggle}
-              label={t('baseplate.syncWithLayout')}
-            />
-            <div className="grid grid-cols-2 gap-x-3 gap-y-2">
-              <GridDimensionStepper
-                label={t('baseplate.gridWidth')}
-                value={effectiveWidth}
-                onChange={(v) => updateParam('baseplateWidth', gridUnits(v))}
-                halfGridMode={halfGridMode}
-                disabled={synced}
-              />
-              <GridDimensionStepper
-                label={t('baseplate.gridDepth')}
-                value={effectiveDepth}
-                onChange={(v) => updateParam('baseplateDepth', gridUnits(v))}
-                halfGridMode={halfGridMode}
-                disabled={synced}
-              />
-            </div>
-            {/* Fractional edge position — shown when the effective dimensions are half-unit */}
-            {(hasFractionalWidth || hasFractionalDepth) && (
-              <div className="space-y-1.5">
-                <div className="text-content-tertiary text-[10px] mb-1">
-                  {t('sidebar.halfUnitEdgePosition')}
-                </div>
-                {hasFractionalWidth && (
-                  <FractionalEdgeToggle
-                    axis="x"
-                    label={t('common.width')}
-                    value={fractionalEdgeX}
-                    onChange={handleFractionalEdgeChange}
-                    startTitle={t('sidebar.halfBinLeft')}
-                    startLabel={t('sidebar.left')}
-                    endTitle={t('sidebar.halfBinRight')}
-                    endLabel={t('sidebar.right')}
-                  />
-                )}
-                {hasFractionalDepth && (
-                  <FractionalEdgeToggle
-                    axis="y"
-                    label={t('common.depth')}
-                    value={fractionalEdgeY}
-                    onChange={handleFractionalEdgeChange}
-                    startTitle={t('sidebar.halfBinBottom')}
-                    startLabel={t('sidebar.bottom')}
-                    endTitle={t('sidebar.halfBinTop')}
-                    endLabel={t('sidebar.top')}
-                  />
-                )}
-              </div>
-            )}
+        <DimensionsSection />
+        <StackPrintSection stackPrint={stackPrint} onChange={setStackPrint} />
+        <BaseSection />
+        <PhysicalUnitsSection />
 
-            {/* mm summary under steppers — matches the layout-mode drawer dimensions pattern.
-                When padding > 0, show a trailing 'incl. padding' note so users see the total
-                is grid + padding, not grid alone. */}
-            <div className="flex flex-wrap items-center justify-center gap-x-1.5 gap-y-0.5 pt-1 text-content-tertiary">
-              <div className="flex items-center gap-1">
-                <RulerIcon size="xs" className="flex-shrink-0" />
-                <EditableDimensions
-                  widthMm={totalWidthMm}
-                  depthMm={totalDepthMm}
-                  minMm={minMm}
-                  maxMm={maxMm}
-                  onCommit={handleDimensionCommit}
-                  aria-label={t('baseplate.editDimensions')}
-                  widthLabel={t('baseplate.editDimensionsWidth')}
-                  depthLabel={t('baseplate.editDimensionsDepth')}
-                  variant="secondary"
-                />
-              </div>
-              {hasPadding && (
-                <span className="text-[11px] italic text-content-tertiary">
-                  {t('baseplate.inclPadding')}
-                </span>
-              )}
-            </div>
-
-            {/* Padding — spatial schematic. Padding composes with every shape,
-                so the controls stay live; a notice explains the shaped case. */}
-            <div className="space-y-2 border-t border-stroke-subtle pt-3">
-              <div className="text-[11px] font-medium uppercase tracking-wide text-content-tertiary">
-                {t('baseplate.padding')}
-              </div>
-              {cornerShaped && (
-                <p className="text-[11px] leading-relaxed text-content-tertiary">
-                  {t('baseplate.cornerShapedPaddingNotice')}
-                </p>
-              )}
-              {freeformShaped && (
-                <p
-                  className={
-                    shapedPaddingComposes
-                      ? 'text-[11px] leading-relaxed text-content-tertiary'
-                      : 'text-[11px] leading-relaxed text-status-warning'
-                  }
-                >
-                  {t(
-                    shapedPaddingComposes
-                      ? 'baseplate.shapedPaddingNotice'
-                      : 'baseplate.shapedPaddingTooLarge'
-                  )}
-                </p>
-              )}
-              <PaddingSchematic
-                baseplateParams={baseplateParams}
-                updateParam={updateParam}
-                updateParams={updateParams}
-              />
-              {hasPadding && (
-                <div className="border-t border-stroke-subtle pt-3">
-                  <FeatureToggle
-                    label={t('baseplate.overTile')}
-                    checked={fillOn}
-                    onChange={() => setMarginFillMode(fillOn ? 'solid' : 'tile')}
-                    disabledReason={
-                      !fillOn && !overTileStatus.canOverTile
-                        ? t('baseplate.overTileTooSmall')
-                        : undefined
-                    }
-                    primaryControls={
-                      <div className="space-y-1.5">
-                        {overTileStatus.canOverTile ? (
-                          <>
-                            <CheckboxRow
-                              label={t('baseplate.preferHalfGrid')}
-                              checked={halfGridOn}
-                              onChange={(checked) =>
-                                setMarginFillMode(checked ? 'halfGrid' : 'tile')
-                              }
-                              indent
-                            />
-                            {halfGridOn && (
-                              <div className="ml-4 flex items-center justify-between gap-2 border-l border-stroke-subtle pl-3">
-                                <span className="text-xs text-content-secondary">
-                                  {t('baseplate.leftoverLabel')}
-                                </span>
-                                <SegmentedControl
-                                  aria-label={t('baseplate.leftoverLabel')}
-                                  size="sm"
-                                  options={[
-                                    { value: 'grid', label: t('baseplate.leftoverGrid') },
-                                    { value: 'solid', label: t('baseplate.leftoverSolid') },
-                                  ]}
-                                  value={leftoverMode}
-                                  onChange={setLeftoverMode}
-                                />
-                              </div>
-                            )}
-                            <div className="space-y-1 text-[11px] leading-relaxed">
-                              <p className="text-content-tertiary">
-                                {t(
-                                  halfGridOn
-                                    ? leftoverMode === 'solid'
-                                      ? 'baseplate.halfGridHintSolid'
-                                      : 'baseplate.halfGridHint'
-                                    : 'baseplate.overTileHint'
-                                )}
-                              </p>
-                              {overTileStatus.tiled.length > 0 && (
-                                <p className="text-content-secondary">
-                                  {t('baseplate.overTileFills', {
-                                    sides: overTileStatus.tiled
-                                      .map((e) => t(e.labelKey))
-                                      .join(', '),
-                                  })}
-                                </p>
-                              )}
-                              {overTileStatus.tooSmall.length > 0 && (
-                                <p className="text-content-tertiary">
-                                  {t('baseplate.overTileKeptSolid', {
-                                    sides: overTileStatus.tooSmall
-                                      .map((e) => t(e.labelKey))
-                                      .join(', '),
-                                  })}
-                                </p>
-                              )}
-                            </div>
-                          </>
-                        ) : (
-                          // Fill is on but no edge can fit a tile right now — keep the
-                          // toggle enabled (so it can be turned off) and explain why.
-                          <p className="text-[11px] leading-relaxed text-content-tertiary">
-                            {t('baseplate.overTileTooSmall')}
-                          </p>
-                        )}
-                      </div>
-                    }
-                  />
-                </div>
-              )}
-              {/* Detached rails have no outline awareness (buildFullParams
-                    forces detachMargins off whenever an outline is active), so
-                    hide the control for every shaped plate, not just corner
-                    cuts. */}
-              {hasPadding && !outlineActive && (
-                <div className="border-t border-stroke-subtle pt-3">
-                  <FeatureToggle
-                    label={t('baseplate.detachMargins')}
-                    checked={detachStored}
-                    onChange={() => updateParam('detachMargins', !detachStored)}
-                    disabledReason={
-                      !detachStored && !canDetach ? t('baseplate.detachMarginsTooSmall') : undefined
-                    }
-                    primaryControls={
-                      // On but nothing meets the threshold → no rails are emitted,
-                      // so say so rather than imply they will.
-                      !canDetach ? (
-                        <p className="text-[11px] leading-relaxed text-content-tertiary">
-                          {t('baseplate.detachMarginsTooSmall')}
-                        </p>
-                      ) : (
-                        <div className="space-y-1">
-                          <CheckboxRow
-                            label={t('baseplate.detachMarginConnector')}
-                            checked={marginConnectorStored && marginConnectorStyleOk}
-                            onChange={(checked) => updateParam('detachMarginConnector', checked)}
-                            disabled={!marginConnectorStyleOk}
-                            indent
-                          />
-                          <p className="text-[11px] leading-relaxed text-content-tertiary pl-6">
-                            {marginConnectorStyleOk
-                              ? t('baseplate.detachMarginConnectorHint')
-                              : t('baseplate.detachMarginConnectorStyle')}
-                          </p>
-                        </div>
-                      )
-                    }
-                  />
-                </div>
-              )}
-            </div>
-          </div>
-        </StickyGroupHeader>
-
-        {/* 2. Stack for printing — experimental. Placed above Base because
-            enabling it strips and hides the Base controls below. */}
-        <StackPrintSection stackPrint={baseplateParams.stackPrint} onChange={setStackPrint} />
-
-        {/* 3. Base — connectors, magnets, corner radius. While stacking, magnets
-            and corner rounding are stripped so they hide, but connectors stay
-            reachable (dovetail styles stack fine); the group only renders when it
-            still has content, so it never shows as an empty collapsible group. */}
-        {(tiling?.isSplit || !stackEnabled) && (
-          <StickyGroupHeader
-            title={t('baseplate.sectionBase')}
-            summary={
-              !stackEnabled && baseplateParams.magnetHoles
-                ? `\u00f8${baseplateParams.magnetDiameter}mm \u00d7 ${baseplateParams.magnetDepth}mm`
-                : undefined
-            }
-          >
-            <div className="space-y-3 px-4 py-3">
-              {tiling?.isSplit && (
-                <ConnectorPicker
-                  value={
-                    // Snap clip is stripped while stacking, so show it as None
-                    // (the effective state); the stored style returns on toggle-off.
-                    stackEnabled && baseplateParams.connectorStyle === 'snapClip'
-                      ? 'none'
-                      : baseplateParams.connectorNubs === true
-                        ? (baseplateParams.connectorStyle ?? 'dovetail')
-                        : 'none'
-                  }
-                  disabledOptions={
-                    stackEnabled
-                      ? { snapClip: t('baseplate.connectors.snapClipNoStack') }
-                      : undefined
-                  }
-                  onChange={(v: ConnectorChoice) => {
-                    if (v === 'none') {
-                      updateParams({ connectorNubs: false, connectorStyle: undefined });
-                      return;
-                    }
-                    // 'dovetail' is the default, stored as undefined.
-                    updateParams({
-                      connectorNubs: true,
-                      connectorStyle:
-                        v === 'dovetailKey' || v === 'snapClip' || v === 'puzzle' ? v : undefined,
-                    });
-                  }}
-                  renderExpanded={(style) =>
-                    style === 'none' ? null : (
-                      <div className="space-y-3">
-                        <SettingsRow
-                          label={t('baseplate.connectorFit.label')}
-                          tooltip={t('baseplate.connectorFit.info')}
-                          unit="mm"
-                        >
-                          <Stepper
-                            size="sm"
-                            value={baseplateParams.connectorFitOffset ?? 0}
-                            onStep={(delta) => {
-                              const next = snapConnectorFitOffset(
-                                (baseplateParams.connectorFitOffset ?? 0) +
-                                  delta * CONNECTOR_FIT_OFFSET_STEP
-                              );
-                              updateParam('connectorFitOffset', next === 0 ? undefined : next);
-                            }}
-                            min={CONNECTOR_FIT_OFFSET_MIN}
-                            max={CONNECTOR_FIT_OFFSET_MAX}
-                            step={CONNECTOR_FIT_OFFSET_STEP}
-                            displayValue={formatConnectorFitOffset(
-                              baseplateParams.connectorFitOffset ?? 0
-                            )}
-                            aria-label={t('baseplate.connectorFit.label')}
-                          />
-                        </SettingsRow>
-                        {(style === 'dovetail' || style === 'puzzle') &&
-                          baseplateParams.preferIdenticalPieces !== true && (
-                            <Checkbox
-                              checked={baseplateParams.invertDovetails === true}
-                              onChange={(checked) =>
-                                updateParam('invertDovetails', checked || undefined)
-                              }
-                              label={t('baseplate.dovetails.invert')}
-                            />
-                          )}
-                        <Checkbox
-                          checked={baseplateParams.preferIdenticalPieces === true}
-                          onChange={(checked) =>
-                            updateParam('preferIdenticalPieces', checked || undefined)
-                          }
-                          label={t('baseplate.preferIdenticalPieces')}
-                        />
-                        <ConnectorSampleButton />
-                        {nozzleSizeMm > NOZZLE_BASELINE && (
-                          <p className="text-[11px] leading-relaxed text-content-tertiary">
-                            {t('baseplate.connectorNozzleNotice', { nozzle: nozzleSizeMm })}
-                          </p>
-                        )}
-                      </div>
-                    )
-                  }
-                />
-              )}
-              {!stackEnabled && (
-                <>
-                  <div className="border-t border-stroke-subtle pt-3">
-                    <FeatureToggle
-                      label={t('baseplate.magnetHoles')}
-                      checked={baseplateParams.magnetHoles}
-                      onChange={() => updateParam('magnetHoles', !baseplateParams.magnetHoles)}
-                      valueSummary={`\u00f8${baseplateParams.magnetDiameter}mm \u00d7 ${baseplateParams.magnetDepth}mm`}
-                    >
-                      <SliderInput
-                        label={t('baseplate.magnetDiameter')}
-                        value={baseplateParams.magnetDiameter}
-                        onChange={(v) => updateParam('magnetDiameter', mm(v))}
-                        min={1}
-                        max={20}
-                        step={0.1}
-                        unit="mm"
-                        info={t('baseplate.magnetDiameterInfo')}
-                      />
-                      <SliderInput
-                        label={t('baseplate.magnetDepth')}
-                        value={baseplateParams.magnetDepth}
-                        onChange={(v) => updateParam('magnetDepth', mm(v))}
-                        min={0.5}
-                        max={10}
-                        step={0.1}
-                        unit="mm"
-                        info={t('baseplate.magnetDepthInfo')}
-                      />
-                    </FeatureToggle>
-                  </div>
-                  <div className="border-t border-stroke-subtle pt-3">
-                    <FeatureToggle
-                      label={t('baseplate.solidFloor')}
-                      // Independent of magnets — the floor is added below the grid
-                      // (and below the magnet layer when present), keeping the
-                      // underside continuous. Thickness is customizable either way.
-                      checked={baseplateParams.solidFloor === true}
-                      onChange={() => updateParam('solidFloor', !baseplateParams.solidFloor)}
-                      primaryControls={
-                        <SliderInput
-                          label={t('baseplate.solidFloorThickness')}
-                          value={baseplateParams.solidFloorThickness ?? SOLID_FLOOR_DEFAULT_MM}
-                          onChange={(v) => updateParam('solidFloorThickness', mm(v))}
-                          min={SOLID_FLOOR_MIN_MM}
-                          max={SOLID_FLOOR_MAX_MM}
-                          step={0.1}
-                          unit="mm"
-                          info={t('baseplate.solidFloorHeightNote')}
-                        />
-                      }
-                    />
-                  </div>
-                  {/* Corner rounding is zeroed whenever an outline is active
-                      (the shape carries its own corners), so show the control
-                      only for an unshaped rectangular plate. */}
-                  {!outlineActive && (
-                    <div className="border-t border-stroke-subtle pt-3">
-                      <CornerRadiusControl
-                        cornerRadius={baseplateParams.cornerRadius}
-                        cornerRadii={baseplateParams.cornerRadii}
-                        // Geometric max (up to a pill shape), floored to the
-                        // slider's 0.5 step. Radii beyond the plain rounding
-                        // limit become a radius-cut outline downstream, which
-                        // trims or drops the sockets the arc consumes.
-                        maxRadius={Math.max(
-                          0,
-                          Math.floor(maxCornerRadiusMm(totalWidthMm, totalDepthMm) * 2) / 2
-                        )}
-                        onUniformChange={(r) => {
-                          updateParam('cornerRadius', mm(r));
-                          updateParam('cornerRadii', undefined);
-                        }}
-                        onPerCornerChange={(radii) => {
-                          updateParam('cornerRadii', {
-                            tl: mm(radii.tl),
-                            tr: mm(radii.tr),
-                            bl: mm(radii.bl),
-                            br: mm(radii.br),
-                          });
-                        }}
-                      />
-                    </div>
-                  )}
-                </>
-              )}
-            </div>
-          </StickyGroupHeader>
-        )}
-
-        {/* 4. Physical Units — grid unit, print bed, nozzle, build height. Mirrors the
-            "Physical Units" section in the bin designer and layout sidebars. */}
-        <StickyGroupHeader
-          title={t('common.physicalUnits')}
-          summary={formatPrintSettingsSummary(
-            gridUnitMm,
-            printBedSize,
-            printBedDepth ?? printBedSize
-          )}
-          expanded={printSettingsExpanded}
-          onExpandedChange={setPrintSettingsExpanded}
-        >
-          <div className="space-y-3 px-4 py-3">
-            <div className="text-xs text-content-secondary space-y-2">
-              <SettingsRow
-                label={t('baseplate.gridUnit')}
-                htmlFor="bp-gridUnit"
-                unit="mm"
-                tooltip={t('baseplate.gridUnitTooltip')}
-              >
-                <DeferredNumberInput
-                  id="bp-gridUnit"
-                  value={gridUnitMm}
-                  onChange={(mm) => useLayoutStore.getState().setGridUnitMm(mm)}
-                  min={1}
-                  max={200}
-                  className="input w-14 py-0.5 px-1 text-xs text-right"
-                />
-              </SettingsRow>
-              {/* Non-square grids get a read-only Y-pitch echo — the toggle and
-                  edit live in the drawer's Physical Units, so the plate stays a
-                  reflection of the layout grid. */}
-              {gridUnitMmY !== gridUnitMm && (
-                <SettingsRow
-                  label={t('baseplate.gridUnitY')}
-                  unit="mm"
-                  tooltip={t('baseplate.gridUnitYTooltip')}
-                >
-                  <span className="px-1 text-xs text-content-secondary tabular-nums">
-                    {gridUnitMmY}
-                  </span>
-                </SettingsRow>
-              )}
-              {/* Progressive disclosure: the anchor only diverges above the
-                  standard 42mm grid (below that 'edge' and 'center' are identical),
-                  so the control stays hidden until then — and its caption reveals
-                  the consequence of the current choice, matching the half-grid
-                  leftover pattern above. */}
-              {gridUnitMm > 42 && (
-                <div className="space-y-1">
-                  <SettingsRow
-                    label={t('baseplate.magnetAnchor')}
-                    tooltip={t('baseplate.magnetAnchorTooltip')}
-                  >
-                    <SegmentedControl<'edge' | 'center'>
-                      aria-label={t('baseplate.magnetAnchor')}
-                      size="sm"
-                      options={[
-                        { value: 'edge', label: t('baseplate.magnetAnchorEdge') },
-                        { value: 'center', label: t('baseplate.magnetAnchorCenter') },
-                      ]}
-                      value={magnetAnchor}
-                      onChange={(value) => useLayoutStore.getState().setMagnetAnchor(value)}
-                    />
-                  </SettingsRow>
-                  <p className="text-[11px] leading-relaxed text-content-tertiary">
-                    {t(
-                      magnetAnchor === 'center'
-                        ? 'baseplate.magnetAnchorHintLegacy'
-                        : 'baseplate.magnetAnchorHintCorners'
-                    )}
-                  </p>
-                </div>
-              )}
-              <HelpTargetMarker id="bp-print-bed-size">
-                <SettingsRow
-                  label={t('baseplate.printBedSize')}
-                  htmlFor="bp-printBedSize"
-                  unit="mm"
-                  tooltip={t('baseplate.printBedTooltip')}
-                >
-                  <PrintBedInput
-                    id="bp-printBedSize"
-                    width={printBedSize}
-                    depth={printBedDepth ?? printBedSize}
-                    onChange={(w, d) => useLayoutStore.getState().setPrintBedSize(w, d)}
-                    variant="compact"
-                  />
-                </SettingsRow>
-              </HelpTargetMarker>
-              <SettingsRow
-                label={t('settings.nozzleSize')}
-                htmlFor="bp-nozzleSize"
-                tooltip={t('baseplate.nozzleSizeTooltip')}
-                unit="mm"
-              >
-                <DeferredNumberInput
-                  id="bp-nozzleSize"
-                  value={nozzleSizeMm}
-                  onChange={handleNozzleChange}
-                  min={PRINT_SETTINGS_CONSTRAINTS.NOZZLE_SIZE_MIN}
-                  max={PRINT_SETTINGS_CONSTRAINTS.NOZZLE_SIZE_MAX}
-                  step={PRINT_SETTINGS_CONSTRAINTS.NOZZLE_SIZE_STEP}
-                  className="input w-14 py-0.5 px-1 text-xs text-right"
-                  aria-label={t('settings.nozzleSize')}
-                />
-              </SettingsRow>
-              <SettingsRow
-                label={t('baseplate.maxPrintHeight')}
-                htmlFor="bp-maxPrintHeight"
-                unit="mm"
-                tooltip={t('baseplate.maxPrintHeightTooltip')}
-              >
-                <DeferredNumberInput
-                  id="bp-maxPrintHeight"
-                  value={maxPrintHeightMm}
-                  onChange={handleMaxHeightChange}
-                  min={PRINT_SETTINGS_CONSTRAINTS.MAX_PRINT_HEIGHT_MIN}
-                  max={PRINT_SETTINGS_CONSTRAINTS.MAX_PRINT_HEIGHT_MAX}
-                  step={PRINT_SETTINGS_CONSTRAINTS.MAX_PRINT_HEIGHT_STEP}
-                  className="input w-14 py-0.5 px-1 text-xs text-right"
-                  aria-label={t('baseplate.maxPrintHeight')}
-                />
-              </SettingsRow>
-            </div>
-          </div>
-        </StickyGroupHeader>
-
-        {/* 4. Split pieces mini-map — only when baseplate is split */}
+        {/* Split pieces mini-map — only when baseplate is split */}
         {tiling?.isSplit && (
           <SplitViewStrip
             tiling={tiling}
@@ -1000,9 +116,4 @@ export function BaseplatePanel() {
       />
     </div>
   );
-}
-/** Print Settings header summary: "{gridUnit}mm · {bed}mm" or "{gridUnit}mm · {w}×{d}mm" for asymmetric beds. */
-function formatPrintSettingsSummary(gridUnitMm: number, bedW: number, bedD: number): string {
-  const bed = bedW === bedD ? `${bedW}mm` : `${bedW}\u00d7${bedD}mm`;
-  return `${gridUnitMm}mm \u00b7 ${bed}`;
 }
